smart_parking_2.1/requirements.txt
# Flask Web框架，版本2.3.3，用于构建Web应用
flask==2.3.3
# Flask的JWT扩展，版本4.5.3，用于实现用户身份验证和授权
flask-jwt-extended==4.5.3
# OpenCV的Python绑定，版本4.7.0.72，用于处理视频和图像
opencv-python==4.7.0.72
# 异步I/O库，版本0.33.3，用于实现异步处理，提高应用性能
eventlet==0.33.3
# 用于加载环境变量的库，版本1.0.0，方便配置应用
python-dotenv==1.0.0
# 密码加密库，版本4.0.1，用于对用户密码进行加密
bcrypt==4.0.1
# 深度学习框架，版本2.6.0+cu126，YOLOv8模型依赖此框架
torch==2.6.0+cu126
# 包含YOLOv8模型的库，版本8.0.154，用于目标检测
ultralytics==8.0.154
# Flask的CORS扩展，用于解决跨域问题
flask-cors==4.0.0